# static/files/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Searchable spatiotemporal metadata catalog",
    "title": "go-stac-server",
    "version": "1.0.0"
  },
  "paths": {
...
    "/collections": {
      "get": {
        "description": "A body of Feature Collections that belong or are used together with additional links.\nRequest may not return the full set of metadata per Feature Collection.",
        "operationId": "getCollections",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Collections"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "the feature collections in the dataset",
        "tags": [
          "Features"
        ]
      },
      "post": {
        "description": "Endpoint.",
        "operationId": "Create_Collection_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/collection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Collection",
        "tags": [
          "Transaction Extension"
        ]
      },
      "put": {
        "description": "Update a collections object",
        "operationId": "updateCollection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/collection"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update Collection",
        "tags": [
          "Transaction Extension"
        ]
      }
    },
    "/collections/{collectionId}": {
      "delete": {
        "description": "Use this method to delete an existing Collection",
        "operationId": "Delete_Collection_collections__collection_id__delete",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/exception"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Collection",
        "tags": [
          "Transaction Extension"
        ]
      },
      "get": {
        "description": "A single Feature Collection for the given `collectionId`.\nRequest this endpoint to get a full list of metadata for the Feature Collection.",
        "operationId": "describeCollection",
        "parameters": [
          {
            "$ref": "#/components/parameters/collectionId"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Collection"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Error"
          },
          "4XX": {
            "$ref": "#/components/responses/Error"
          },
          "5XX": {
            "$ref": "#/components/responses/Error"
          }
        },
        "summary": "describe the feature collection with id `collectionId`",
        "tags": [
          "Features"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/stac/v1"
    }
  ],
...
  "components": {
    "parameters": {
...
      "collectionId": {
        "description": "local identifier of a collection",
        "in": "path",
        "name": "collectionId",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
...
      "Collection": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/collection"
            }
          }
        },
        "description": "Information about the feature collection with id `collectionId`.\n\nThe response contains a link to the items in the collection\n(path `/collections/{collectionId}/items`, link relation `items`)\nas well as key information about the collection. This information\nincludes:\n\n* A local identifier for the collection that is unique for the dataset;\n* A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);\n* An optional title and description for the collection;\n* An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;\n* An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is \"feature\")."
      },
      "Collections": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/collections"
            }
          }
        },
        "description": "The feature collections shared by this API.\n\nThe dataset is organized as one or more feature collections. This resource\nprovides information about and access to the collections.\n\nThe response contains the list of collections. For each collection, a link\nto the items in the collection (path `/collections/{collectionId}/items`,\nlink relation `items`) as well as key information about the collection.\nThis information includes:\n\n* A local identifier for the collection that is unique for the dataset;\n* A list of coordinate reference systems (CRS) in which geometries may be returned by the server. The first CRS is the default coordinate reference system (the default is always WGS 84 with axis order longitude/latitude);\n* An optional title and description for the collection;\n* An optional extent that can be used to provide an indication of the spatial and temporal extent of the collection - typically derived from the data;\n* An optional indicator about the type of the items in the collection (the default value, if the indicator is not provided, is \"feature\")."
      },
...
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/exception"
            }
          }
        },
        "description": "An error occurred."
      },
...
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/exception"
            }
          }
        },
        "description": "The requested URI was not found."
      },
...
    },
    "schemas": {
...
      "collection": {
        "examples": [
          {
            "description": "Sentinel-2 is a wide-swath, high-resolution, multi-spectral\nimaging mission...\n",
            "extent": {
              "spatial": {
                "bbox": [
                  [
                    -180,
                    -56,
                    180,
                    83
                  ]
                ]
              },
              "temporal": {
                "interval": [
                  [
                    "2015-06-23T00:00:00Z",
                    "2019-07-10T13:44:56Z"
                  ]
                ]
              }
            },
            "id": "Sentinel-2",
            "keywords": [
              "copernicus",
              "esa",
              "eu",
              "msi",
              "radiance",
              "sentinel"
            ],
            "license": "proprietary",
            "links": [
              {
                "href": "http://stac.server/api/stac/v1/collections/Sentinel-2",
                "rel": "self",
                "type": "application/json"
              },
              {
                "href": "http://stac.server/api/stac/v1/collections",
                "rel": "root",
                "type": "application/json"
              },
              {
                "href": "https://scihub.copernicus.eu/twiki/pub/SciHubWebPortal/TermsConditions/Sentinel_Data_Terms_and_Conditions.pdf",
                "rel": "license",
                "title": "Legal notice on the use of Copernicus Sentinel Data and Service Information",
                "type": "application/pdf"
              }
            ],
            "providers": [
              {
                "name": "ESA",
                "roles": [
                  "producer",
                  "licensor"
                ],
                "url": "https://sentinel.esa.int/web/sentinel/user-guides/sentinel-2-msi"
              }
            ],
            "stac_extensions": [],
            "stac_version": "1.0.0",
            "summaries": {
              "constellation": [
                "sentinel-2"
              ],
              "datetime": {
                "max": "2019-07-10T13:44:56Z",
                "min": "2015-06-23T00:00:00Z"
              },
              "eo:bands": [
                [
                  {
                    "center_wavelength": 4.439,
                    "common_name": "coastal",
                    "name": "B1"
                  },
                  {
                    "center_wavelength": 4.966,
                    "common_name": "blue",
                    "name": "B2"
                  },
                  {
                    "center_wavelength": 5.6,
                    "common_name": "green",
                    "name": "B3"
                  },
                  {
                    "center_wavelength": 6.645,
                    "common_name": "red",
                    "name": "B4"
                  },
                  {
                    "center_wavelength": 7.039,
                    "name": "B5"
                  },
                  {
                    "center_wavelength": 7.402,
                    "name": "B6"
                  },
                  {
                    "center_wavelength": 7.825,
                    "name": "B7"
                  },
                  {
                    "center_wavelength": 8.351,
                    "common_name": "nir",
                    "name": "B8"
                  },
                  {
                    "center_wavelength": 8.648,
                    "name": "B8A"
                  },
                  {
                    "center_wavelength": 9.45,
                    "name": "B9"
                  },
                  {
                    "center_wavelength": 1.3735,
                    "name": "B10"
                  },
                  {
                    "center_wavelength": 1.6137,
                    "common_name": "swir16",
                    "name": "B11"
                  },
                  {
                    "center_wavelength": 2.2024,
                    "common_name": "swir22",
                    "name": "B12"
                  }
                ]
              ],
              "eo:gsd": [
                10,
                30,
                60
              ],
              "instruments": [
                "msi"
              ],
              "platform": [
                "sentinel-2a",
                "sentinel-2b"
              ],
              "sci:citation": [
                "Copernicus Sentinel data [Year]"
              ],
              "view:off_nadir": {
                "max": 100,
                "min": 0
              },
              "view:sun_elevation": {
                "max": 89.9,
                "min": 6.78
              }
            },
            "title": "Sentinel-2 MSI: MultiSpectral Instrument, Level-1C",
            "type": "Collection"
          }
        ],
        "properties": {
          "description": {
            "description": "Detailed multi-line description to fully explain the catalog or collection.\n[CommonMark 0.29](http://commonmark.org/) syntax MAY be used for rich text representation.",
            "type": "string"
          },
          "extent": {
            "$ref": "#/components/schemas/extent"
          },
          "id": {
            "description": "identifier of the collection used, for example, in URIs",
            "type": "string"
          },
          "keywords": {
            "description": "List of keywords describing the collection.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "license": {
            "$ref": "#/components/schemas/license"
          },
          "links": {
            "$ref": "#/components/schemas/links"
          },
          "providers": {
            "$ref": "#/components/schemas/providers"
          },
          "stac_extensions": {
            "items": {
              "anyOf": [
                {
                  "format": "uri",
                  "title": "Reference to a JSON Schema",
                  "type": "string"
                },
                {
                  "title": "Reference to a core extension",
                  "type": "string"
                }
              ]
            },
            "title": "STAC extensions",
            "type": "array",
            "uniqueItems": true
          },
          "stac_version": {
            "examples": [
              "1.0.0"
            ],
            "title": "STAC version",
            "type": "string"
          },
          "summaries": {
            "additionalProperties": {
              "oneOf": [
                {
                  "items": {
                    "description": "A value of any type."
                  },
                  "title": "Set of values",
                  "type": "array"
                },
                {
                  "description": "By default, only ranges with a minimum and a maximum value can\nbe specified. Ranges can be specified for ordinal values only,\nwhich means they need to have a rank order. Therefore, ranges\ncan only be specified for numbers and some special types of\nstrings. Examples: grades (A to F), dates or times.\nImplementors are free to add other derived statistical values\nto the object, for example `mean` or `stddev`.",
                  "properties": {
                    "max": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    },
                    "min": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "number"
                        }
                      ]
                    }
                  },
                  "required": [
                    "min",
                    "max"
                  ],
                  "title": "Statistics",
                  "type": "object"
                }
              ]
            },
            "description": "Summaries are either a unique set of all available values *or*\nstatistics. Statistics by default only specify the range (minimum\nand maximum values), but can optionally be accompanied by additional\nstatistical values. The range can specify the potential range of\nvalues, but it is recommended to be as precise as possible. The set\nof values must contain at least one element and it is strongly\nrecommended to list all values. It is recommended to list as many\nproperties as reasonable so that consumers get a full overview of\nthe Collection. Properties that are covered by the Collection\nspecification (e.g. `providers` and `license`) may not be repeated\nin the summaries.",
            "type": "object"
          },
          "title": {
            "description": "human readable title of the collection",
            "type": "string"
          },
          "type": {
            "enum": [
              "Collection"
            ],
            "type": "string"
          }
        },
        "required": [
          "stac_version",
          "type",
          "id",
          "description",
          "license",
          "extent",
          "links"
        ],
        "type": "object"
      },
      "collections": {
        "properties": {
          "collections": {
            "items": {
              "$ref": "#/components/schemas/collection"
            },
            "type": "array"
          },
          "links": {
            "$ref": "#/components/schemas/links"
          }
        },
        "required": [
          "links",
          "collections"
        ],
        "type": "object"
      },
...
      "exception": {
        "description": "Information about the exception: an error code plus an optional description.",
        "properties": {
          "code": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
      "extent": {
        "description": "The extent of the features in the collection. In the Core only spatial and temporal\nextents are specified. Extensions may add additional members to represent other\nextents, for example, thermal or pressure ranges.\n\nThe first item in the array describes the overall extent of\nthe data. All subsequent items describe more precise extents,\ne.g., to identify clusters of data.\nClients only interested in the overall extent will only need to\naccess the first item in each array.",
        "properties": {
          "spatial": {
            "description": "The spatial extent of the features in the collection.",
            "properties": {
              "bbox": {
                "description": "One or more bounding boxes that describe the spatial extent of the dataset.\n\nThe first bounding box describes the overall spatial\nextent of the data. All subsequent bounding boxes describe\nmore precise bounding boxes, e.g., to identify clusters of data.\nClients only interested in the overall spatial extent will\nonly need to access the first item in each array.",
                "items": {
                  "description": "Each bounding box is provided as four or six numbers, depending on\nwhether the coordinate reference system includes a vertical axis\n(height or depth):\n\n* Lower left corner, coordinate axis 1\n* Lower left corner, coordinate axis 2\n* Minimum value, coordinate axis 3 (optional)\n* Upper right corner, coordinate axis 1\n* Upper right corner, coordinate axis 2\n* Maximum value, coordinate axis 3 (optional)\n\nThe coordinate reference system of the values is WGS 84 longitude/latitude\n(http://www.opengis.net/def/crs/OGC/1.3/CRS84) unless a different coordinate\nreference system is specified in `crs`.\n\nFor WGS 84 longitude/latitude the values are in most cases the sequence of\nminimum longitude, minimum latitude, maximum longitude and maximum latitude.\nHowever, in cases where the box spans the antimeridian the first value\n(west-most box edge) is larger than the third value (east-most box edge).\n\nIf the vertical axis is included, the third and the sixth number are\nthe bottom and the top of the 3-dimensional bounding box.\n\nIf a feature has multiple spatial geometry properties, it is the decision of the\nserver whether only a single spatial geometry property is used to determine\nthe extent or all relevant geometries.",
                  "examples": [
                    -180,
                    -90,
                    180,
                    90
                  ],
                  "items": {
                    "type": "number"
                  },
                  "maxItems": 6,
                  "minItems": 4,
                  "type": "array"
                },
                "minItems": 1,
                "type": "array"
              },
              "crs": {
                "default": "http://www.opengis.net/def/crs/OGC/1.3/CRS84",
                "description": "Coordinate reference system of the coordinates in the spatial extent\n(property `bbox`). The default reference system is WGS 84 longitude/latitude.\nIn the Core this is the only supported coordinate reference system.\nExtensions may support additional coordinate reference systems and add\nadditional enum values.",
                "enum": [
                  "http://www.opengis.net/def/crs/OGC/1.3/CRS84"
                ],
                "type": "string"
              }
            },
            "required": [
              "bbox"
            ],
            "type": "object"
          },
          "temporal": {
            "description": "The temporal extent of the features in the collection.",
            "properties": {
              "interval": {
                "description": "One or more time intervals that describe the temporal extent of the dataset.\n\nThe first time interval describes the overall\ntemporal extent of the data. All subsequent time intervals describe\nmore precise time intervals, e.g., to identify clusters of data.\nClients only interested in the overall extent will only need\nto access the first item in each array.",
                "items": {
                  "description": "Begin and end times of the time interval. The timestamps\nare in the coordinate reference system specified in `trs`. By default\nthis is the Gregorian calendar.\n\nThe value `null` is supported and indicates an open time interval.",
                  "examples": [
                    "2011-11-11T12:22:11Z",
                    null
                  ],
                  "items": {
                    "format": "date-time",
                    "nullable": true,
                    "type": "string"
                  },
                  "maxItems": 2,
                  "minItems": 2,
                  "type": "array"
                },
                "minItems": 1,
                "type": "array"
              },
              "trs": {
                "default": "http://www.opengis.net/def/uom/ISO-8601/0/Gregorian",
                "description": "Coordinate reference system of the coordinates in the temporal extent\n(property `interval`). The default reference system is the Gregorian calendar.\nIn the Core this is the only supported temporal reference system.\nExtensions may support additional temporal reference systems and add\nadditional enum values.",
                "enum": [
                  "http://www.opengis.net/def/uom/ISO-8601/0/Gregorian"
                ],
                "type": "string"
              }
            },
            "required": [
              "interval"
            ],
            "type": "object"
          }
        },
        "required": [
          "spatial",
          "temporal"
        ],
        "type": "object"
      },
...
      "license": {
        "description": "License(s) of the data as a SPDX\n[License identifier](https://spdx.org/licenses/). Alternatively, use\n`proprietary` if the license is not on the SPDX license list or\n`various` if multiple licenses apply. In these two cases links to the\nlicense texts SHOULD be added, see the `license` link relation type.\n\nNon-SPDX licenses SHOULD add a link to the license text with the\n`license` relation in the links section. The license text MUST NOT be\nprovided as a value of this field. If there is no public license URL\navailable, it is RECOMMENDED to host the license text and\nlink to it.",
        "examples": [
          "Apache-2.0"
        ],
        "type": "string"
      },
...
      "links": {
        "items": {
          "$ref": "#/components/schemas/link"
        },
        "type": "array"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}